refactor: 사용하는 Store DTO 위치 변경
