+++ src/main/java/com/burntoburn/easyshift/controller/StoreController.java
@@ -1,7 +1,10 @@
 package com.burntoburn.easyshift.controller;
 
 import com.burntoburn.easyshift.common.response.ApiResponse;
-import com.burntoburn.easyshift.dto.store.*;
+import com.burntoburn.easyshift.dto.store.use.StoreCreateRequest;
+import com.burntoburn.easyshift.dto.store.use.StoreCreateResponse;
+import com.burntoburn.easyshift.dto.store.use.StoreInfoResponse;
+import com.burntoburn.easyshift.dto.store.use.UserStoresResponse;
 import com.burntoburn.easyshift.service.store.StoreServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
