+++ src/main/java/com/burntoburn/easyshift/dto/store/use/StoreCreateResponse.java
@@ -1,4 +1,4 @@
-package com.burntoburn.easyshift.dto.store;
+package com.burntoburn.easyshift.dto.store.use;
 
 
 import lombok.AllArgsConstructor;
