+++ src/main/java/com/burntoburn/easyshift/dto/store/use/UserStoresResponse.java
@@ -1,5 +1,6 @@
-package com.burntoburn.easyshift.dto.store;
+package com.burntoburn.easyshift.dto.store.use;
 
+import com.burntoburn.easyshift.dto.store.StoreResponse;
 import com.burntoburn.easyshift.entity.store.Store;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
